requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
pypdf==4.1.0
langchain==0.1.13
langchain-google-genai==1.0.1
langchain-community==0.0.29
langchain-core==0.1.33
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1